First working version of chunks
