+++ src/js/PlayerFactory.js
@@ -4,8 +4,8 @@ const PartyStatus = require('./ui/PartyStatus');
 const Constants = require('./Constants');
 
 const PlayerFactory = {
-	buildPlayer: function(game, level, x, y, z){
-		const mob = NPCFactory.buildNPC(game, "avatar", level, x, y, x);
+	buildPlayer: function(game, world, x, y, z){
+		const mob = NPCFactory.buildNPC(game, "avatar", world, x, y, z);
 		mob.canStartDialog = true;
 		PlayerStateMachine.inventory = mob.inventory;
 		mob.name = "Avatar";
