+++ src/js/model/World.class.js
@@ -1,10 +1,15 @@
 const circular = require('circular-functions');
+const ChunkLoader = require('../ChunkLoader');
 const SkyBox = require('../SkyBox');
 const Timer = require('../Timer');
 
-function World () {
+function World (config) {
     this._c = circular.register('World');
 	this.currentMinuteOfDay = 0;
+	this.mobs = [];
+	this.chunkSize = config.chunkSize;
+	this.width = config.chunkSize * config.chunksWidth;
+	this.height =  config.chunkSize * config.chunksHeight;
 }
 
 circular.registerClass('World', World);
@@ -39,7 +44,7 @@ World.prototype = {
 	},
 
 	hourlyNotification () {
-		OAX6.UI.player.level.hourly();
+		OAX6.UI.player.world.hourly();
 	},
 
 	getHourOfDay () {
@@ -61,7 +66,296 @@ World.prototype = {
 		} else {
 			OAX6.UI.setFogColor(0, 0);
 		}
+	},
+
+	getMobAt: function(x, y, z){
+		// TODO: Replace for this.mobs.find
+		for (var i=0,mob;mob=this.mobs[i];i++) {
+			if (mob.x == x && mob.y == y && mob.z == z) {
+				return this.mobs[i];
+			}
+		}
+		return false;
+	},
+	addMob: function(mob){
+		this.mobs.push(mob);
+	},
+	removeMob: function(mob){
+		this.mobs.splice(this.mobs.indexOf(mob), 1);
+	},
+	/**
+	 * When in combat mode, determines the next mob to act, and calls its `act` function
+	 */
+	 actNext: function(){
+		if (this.destroyed) {
+			// This world is no longer active, so stop processing the queue
+			return;
+		}
+		if (PlayerStateMachine.isPartyDead()) { return; }
+		if (PlayerStateMachine.state !== PlayerStateMachine.COMBAT) {
+			// There was a state transition and the world is no longer
+			// in charge of handling the mob actions
+			return;
+		}
+
+		const nextActor = this.mobs[this.currentTurnCounter++];
+		if (this.currentTurnCounter >= this.mobs.length) {
+			this.currentTurnCounter = 0;
+		}
+
+		if (!nextActor || nextActor.dead || !this.isInCombat(nextActor)){
+			return this.actNext();
+		}
+		OAX6.UI.showMessage(nextActor.getBattleDescription()+":");
+		if (nextActor.isPartyMember()){
+			OAX6.UI.locateMarker(nextActor);
+			nextActor.act();
+			// Player will take its time, then call actNext himself
+			// via the PlayerStateMachine
+		} else {
+			Timer.delay(250)
+			.then(()=>{
+				OAX6.UI.hideMarker();
+			});
+			nextActor.act()
+			.then(()=>{
+				if (OAX6.UI.activeMob !== nextActor){
+					OAX6.UI.hideMarker();
+					this.actNext();
+				}
+			});
+		} 
+	},
+	isMobActive: function(){
+		return this.mobs.find(m=>m.executingAction === true);
+	},
+	getMobsOfAlignment: function(alignment){
+		return this.mobs.filter(m=>m.alignment === alignment);
+	},
+	getCloserMobTo: function(x, y, z, alignment){
+		const mobs = this.mobs.filter(m => m.alignment === alignment && m.z == z);
+		if (mobs.length > 0){
+			const sorted = mobs.sort((a,b)=>Geo.flatDist(x,y,a.x,a.y)-Geo.flatDist(x,y,b.x,b.y));
+			return sorted[0];
+		}
+		return false;
+	},
+	/**
+	 * Finds a "long" path to a far location, this is used by NPC schedules AI to walk to their destinations,
+	 * It ignores unlocked doors as blocking spaces, because the NPC can and will open them.
+	 * 
+	 * @param {*} to 
+	 * @param {*} from 
+	 * @param {*} z 
+	 * @param {*} includeMobsOfAlignment 
+	 */
+	findLongPath: function(to, from, z, includeMobsOfAlignment){
+		const gridClone = this._getGridWithMobsBlocked(to, z, includeMobsOfAlignment);
+		this.doors.forEach(door => {
+			if (door.z != z) {
+				return;
+			}
+			if (door.def.fixed) {
+				return;
+			}
+			if (!door.isLocked()) {
+				gridClone.setWalkableAt(door.x, door.y, true);
+			}
+		});
+		// Prevent walking thru stairs
+		this.objects.filter(o => o.z === z && o.type == 'Stairs').forEach(stairs => gridClone.setWalkableAt(stairs.x, stairs.y, false));
+		return this.findPath(from, to, z, gridClone, true)
+	},
+
+	findPathThruMobs: function(to, from, z, includeMobsOfAlignment){
+		const gridClone = this._getGridWithMobsBlocked(to, z, includeMobsOfAlignment);
+		return this.findPath(from, to, z, gridClone)
+	},
+	_getGridWithMobsBlocked(to, z, includeMobsOfAlignment) {
+		const gridClone = this.pfGrids[z].clone();
+		
+		const includeAlignments = [Constants.Alignments.NEUTRAL, includeMobsOfAlignment];
+		includeAlignments.forEach((alignment) => {
+			if (alignment){
+				const mobs = this.getMobsOfAlignment(alignment);
+				mobs.forEach(m=> {
+					if (!(m.x === to.x && m.y === to.y)){
+						gridClone.setWalkableAt(m.x, m.y, false);
+					}
+				});
+			}
+		});
+		return gridClone;
+	},
+	findPath(from, to, z, grid, openEnded) {
+		if (!grid) {
+			grid = this.pfGrids[z].clone();
+		}
+		if (openEnded) {
+			grid.setWalkableAt(from.x, from.y, true);
+			grid.setWalkableAt(to.x, to.y, true);
+		}
+		//TODO: Single finder object?
+		const finder = new PF.AStarFinder({
+		    allowDiagonal: true,
+    		dontCrossCorners: false,
+			diagonalMovement: PF.DiagonalMovement.Always
+		});
+		const path = finder.findPath(from.x, from.y, to.x, to.y, grid);
+		if (path.length == 0){
+			return {dx:0, dy:0};
+		}
+		return {
+			dx: Math.sign(path[1][0]-from.x),
+			dy: Math.sign(path[1][1]-from.y)
+		};
+	},
+	sortForCombat: function(playerGoesFirst){
+		this.currentTurnCounter = 0;
+		this.mobs = this.mobs.sort((a,b)=>a.speed.current - b.speed.current);
+		
+		const player = OAX6.UI.player;
+		if (!player.dead && playerGoesFirst){
+			this.removeMob(player);
+			this.mobs.unshift(player);
+		}
+	},
+	isSafeAround: function(x, y, alignment, panickedAreDangerous){
+		const dangerousMob = this.mobs.find((m) => {
+			const isDangerous = m.alignment !== Constants.Alignments.NEUTRAL &&
+				m.alignment !== alignment &&
+				Geo.flatDist(x,y,m.x,m.y) < COMBAT_DISTANCE;
+			if (isDangerous && !panickedAreDangerous && m.isPanicked) {
+				return false;
+			}
+			return isDangerous;
+		});
+		return !dangerousMob;
+	},
+	activateAll: function(){
+		this.mobs.forEach(m=>m.activate());	
+	},
+	isInCombat: function(mob) {
+		if (mob.hasBeenAttacked) {
+			return true;
+		}
+		const bbox = PlayerStateMachine.getPartyBoundingBox();
+		const width = bbox.x2 - bbox.x1;
+		const height = bbox.y2 - bbox.y1;
+		const x = bbox.x1 + width / 2;
+		const y = bbox.y1 + height / 2;
+
+		const dx = Math.abs(mob.x - x) - width / 2;
+		const dy = Math.abs(mob.y - y) - height / 2;
+
+		const dist = Geo.flatDist(0, 0, dx, dy);
+
+		return (dist < COMBAT_DISTANCE);
+	},
+	// Changes the intent of all enemy mobs with 'waitCommand' intent to 'seekPlayer'
+	activateHostile() {
+		const mobs = this.getMobsOfAlignment(Constants.Alignments.ENEMY);
+		mobs.forEach(m=> {
+			if (m.intent === 'waitCommand') {
+				m.intent = 'seekPlayer';
+			}
+		});
+	},
+	destroy() {
+		// Silently kill all mobs except player party.
+		this.mobs.forEach(m => {
+			if (!m.isPartyMember()) {
+				m.dead = true;
+			}
+		});
+		this.destroyed = true;
+		this.mobs = null;	
+	},
+
+	hourly() {
+		OAX6.UI.player.party.forEach(function(partyMember) {
+			partyMember.hunger.increase(1);
+		});
+		// Check for NPCs that should wake up
+		this.mobs.filter(m => m.isAsleep).forEach(m => m.hourlyWakeupCheck());
+		OAX6.UI.updateFOV(); // Because the new hour may have less visibility
+	},
+
+	isLineClear: function (xa, ya, xb, yb, z) {
+		return !Line.checkInLine(xa, ya, xb, yb, (x, y) => this.isSolid(x, y, z));
+	},
+
+	isSolid: function(x, y, z){
+		return this._getFromChunk(x, y, z, 'isSolid');
+	},
+	isOpaque: function(x, y, z){
+		return this._getFromChunk(x, y, z, 'isOpaque');
+	},
+	getDoorAt: function(x, y, z) {
+		return this._getFromChunk(x, y, z, 'getDoorAt');
+	},
+	getObjectAt: function(x, y, z) {
+		return this._getFromChunk(x, y, z, 'getObjectAt');
+	},
+	getItemAt: function(x, y, z) {
+		return this._getFromChunk(x, y, z, 'getItemAt');
+	},
+	_getFromChunk(x, y, z, functionName) {
+		const chunkX = Math.floor(x / this.chunkSize);
+		const chunkY = Math.floor(y / this.chunkSize);
+		const chunk = this._getChunk(chunkX, chunkY);
+		const chunkInX = x % this.chunkSize;
+		const chunkInY = y % this.chunkSize;
+		return chunk[functionName](chunkInX, chunkInY, z);
+	},
+	removeItem: function(item) {
+		const chunk = item.container; // TODO: Can we confirm this is a chunk?
+		return chunk.removeItem(item);
+	},
+	addItem: function(item, x, y, z) {
+		const chunkX = Math.floor(x / this.chunkSize);
+		const chunkY = Math.floor(y / this.chunkSize);
+		const chunk = this._getChunk(chunkX, chunkY);
+		const chunkInX = x % this.chunkSize;
+		const chunkInY = y % this.chunkSize;
+		return chunk.addItem(item, chunkInX, chunkInY, z);
+	},
+
+	canMoveFrom: function(x, y, z, dx, dy) {
+		if (!this.isValidCoordinate(x + dx, y + dy)) {
+			return false;
+		}
+		if (dx === 0 || dy === 0){
+			return !this.isSolid(x + dx, y + dy, z);
+		}
+		if (this.isSolid(x + dx, y, z) && this.isSolid(x, y + dy, z)){
+			return false;
+		}
+		return !this.isSolid(x + dx, y + dy, z);
+	},
+	isValidCoordinate(x, y) {
+		if (x < 0 || x > this.width || y < 0 || y > this.height) {
+			return false;
+		}
+		return true;
+	},
+	/**
+	 * WARNING: This function should be use sparsely because it breaks the
+	 * chunk abstraction.
+	 * 
+	 * @param {*} x
+	 * @param {*} y
+	 */
+	getCurrentChunk(x, y) {
+		const chunkX = Math.floor(x / this.chunkSize);
+		const chunkY = Math.floor(y / this.chunkSize);
+		return this._getChunk(chunkX, chunkY);
+	},
+
+	_getChunk(x, y) {
+		return ChunkLoader.getChunk(x, y, this);
 	}
+
 }
 
 module.exports = World;
